+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [14.6.0-dev.1] - 2019-12-01
+
 - change init interactive to be true by default
 - compress ssh args before sending